+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-08-18 13:37:44
-// Commit: feat(test): add bi-directional event bus handler for test
+// Updated at: 2026-08-18 13:45:49
+// Commit: feat(test): implement shimmer loading skeleton in test
 
 class testComponentState {
-  final String commitRef = '3695';
-  final String lastMessage = 'feat(test): add bi-directional event bus handler for test';
-  final DateTime timestamp = DateTime.parse('2026-08-18T13:37:44+05:00');
+  final String commitRef = '3696';
+  final String lastMessage = 'feat(test): implement shimmer loading skeleton in test';
+  final DateTime timestamp = DateTime.parse('2026-08-18T13:45:49+05:00');
 }