+++ CHANGELOG.md
@@ -3697,3 +3697,4 @@
 - [2026-08-18 13:33] test(theme): add unit tests for theme business logic
 - [2026-08-18 13:34] feat(network): add async network stream provider
 - [2026-08-18 13:37] feat(test): add bi-directional event bus handler for test
+- [2026-08-18 13:45] feat(test): implement shimmer loading skeleton in test